Return existing ActionResult on Update if it exists.
Mark ourselves as not updateable. It's not clear from the docs whether we should do this (since you can't update an existing one)
or not (since you can call UpdateActionResult to upload a new one). Not that it matters since we only have one client which
doesn't care about this field anyway.