Short-circuit Bytestream.Read requests at the end of the blob.
This can come up if the connection is broken between reading the last section of the blob and the client getting the trailers to terminate the RPC; they re-issue a request which
should get terminated immediately, but we get an error from GCS because we're outside the range of the blob we're reading.
Ideally the SDK would actually recognise this and skip the request, but we can tackle that later.