Add metrics to measure latency with redis in mettle worker. (#296)
We are planning on making elan read from redis as at the moment it is costing us >100£ on normal days to read from the bucket (can reach 300£/day on toolchain upgrades). However we're seeing a lot of context deadline exceeded in the logs regarding redis, so we need to fix those before doing the change to elan. This PR is for us to start monitoring the latency of redis commands to help us understand the impact of the fixes we'll add.